New post: Blogging With Raspberry Pi
There are a million ways to publish a blog, but I happen to think it's
super cool that I can host the whole enchilada on a 5-watt, $25 computer
sitting down in my family room.

In this post I show I got my blog running on a Pi with lighttpd and a
publishing tool called Pelican.